Code to display biometric images for preview form added
