+++ client/app/src/main/java/io/mosip/registration/app/util/ClientConstants.java
@@ -6,6 +6,7 @@ public class ClientConstants {
     public static final String REQUIRED_FIELD_LABEL_TEMPLATE = "%s <font color='#FF0000'>*</font>";
     public static final String FIELD_LABEL_TEMPLATE = "%s";
     public static final String LABEL_SEPARATOR = " / ";
+    public static final String R_ID = "rId";
 
     //Settings preference
     public static final String WIFI_ONLY_PREFERENCE = "wifi_only";
