+++ client/clientmanager/build.gradle
@@ -99,6 +99,7 @@ dependencies {
     implementation 'com.fasterxml.jackson.core:jackson-databind:2.13.2.2'
     implementation 'com.fasterxml.jackson.core:jackson-annotations:2.13.2'
 
+    implementation 'org.apache.velocity:velocity:1.7'
     implementation 'org.mvel:mvel2:2.4.14.Final'
 
     implementation('com.tom-roush:pdfbox-android:2.0.20.0') {
